refactor: 일부 파일 디렉토리 수정 - template.js + exchange.js를 utils 폴더로 이동 (MVC를 도와주는 역할이기 때문에)
